fix: 選定候補の要求枚数保証を追加する
